requirements.txt
pillow
argparse
ksupk
numpy
tqdm
gradio
pyyaml
pdf2image
